[ALPS05434742] memory: Fix a bug of addr2dram
The addr2dram function needs to check the integrity of the input
address.  If the address is over the maximal DRAM size, it should
return failure directly.

The input address is given from the CPUs point of view(1GB register
offset + DRAM offset). On checking out-of-DRAM-max-size, the input
address should be adjusted (-1GB offset) first.

This patch fixes a software bug. Adjust the input address before
compare it with the maximal DRAM size.

MTK-Commit-Id: 7ae63384c6022419184d550141a318d760391858

Change-Id: I92cdff6748ab3dec71d662bac9478dce41e59a3e
Signed-off-by: mtk21827 <joanne.wang@mediatek.com>
CR-Id: ALPS05434742
Feature: External Memory Interface(emi)