[ALPS06278376] net: split out functions related to registering inflight socket files
commit

We need this functionality for the io_uring file registration, but
we cannot rely on it since CONFIG_UNIX can be modular. Move the helpers
to a separate file, thats always builtin to the kernel if CONFIG_UNIX is
m/y.

No functional changes in this patch, just moving code around.

MTK-Commit-Id: fe368d7c80e39926ef48401820e414d40bb725f6

Reviewed-by: Hannes Reinecke <hare@suse.com>
Acked-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Jens Axboe <axboe@kernel.dk>
[ backported to older kernels to get access to unix_gc_lock - gregkh ]
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Feature: Google Security Patch
CR-Id: ALPS06278376
Change-Id: Id0c635cc342cd2e82fddaf93edf30e5e33ff0d53