[ALPS06316203] binder: Disable debug KE for netd
This KE is to debug netd binder full issue. After a period of SQC
we find the issue root cause and catch some issue APKs. The debug KE
finishes its task successfully.
So now we plan to comment the KE and just leave a warning log to
monitor in case of issue from customer.

MTK-Commit-Id: c923f34469be5c8a25d601dd888b5828cc34a63d

Signed-off-by: lena wang <lena.wang@mediatek.com>
CR-Id: ALPS06316203
Feature: [Module]Binder
Change-Id: Ica45c701ca846c2baa323e9e26ca1b5e282df24b