[ALPS06430519] GPU: secure patch GPUCORE-30405
[PATCH] GPUCORE-30405 Prevent use-after-free after region tracker out-of-memory

Convert kbase_remove_va_region() to not be able to fail, by accepting
that if we cannot reallocate a free region in the gap left after
removing a region, then there is nothing we can do and should just leave
the gap.

Instead, completely unlink the removed region regardless so that it
cannot be in use in case the caller kfree()s it. An allocation attempted
into that gap (SAME_VA allocations only) will fail gracefully.

This fixes errata GPUSWERRATA-1207(Bifrost) and GPUSWERRATA-1216(Valhall)

MTK-Commit-Id: f27b09416048b37159dbef24f787259a6b9143da

Signed-off-by: GuanWen Chen <gw.chen@mediatek.com>
CR-Id: ALPS06430519
Feature: GPU ARM(gpu_arm)
Change-Id: Ic68c0942fbb0bcae63a4acac81f64d5cd6bb5101